### 2021-08-09 #3576 @timtroendle
:wave: @noah80, @mewilhel & @robinroche, would any of you be willing to review this submission for JOSS?
