### 2021-08-10 #3596 @whedon
[REVIEW]: Nempy: A Python package for modelling the Australian National Electricity Market dispatch procedure
**Submitting author:** @nick-gorman (<a href="http://orcid.org/0000-0002-6505-761X">Nicholas Gorman</a>)
**Repository:** <a href="https://github.com/UNSW-CEEM/nempy" target ="_blank">https://github.com/UNSW-CEEM/nempy</a>
**Version:** v1.1.2
