### 2021-08-10 #3596 @timtroendle
@noah80, @robinroche, I'll add a reminder for both of you to notify you when half of the six weeks has elapsed.
