### 2021-09-01 97e30f2
get market dispatch to return time taken by solver
