### 2021-09-20 #3596 @robinroche
@timtroendle
@nick-gorman

Thanks for this interesting submission. I installed the package and ran the examples without trouble.

After reviewing the submission, here are some suggestions for improvement:
1. While I understand the general purpose of the code, additional information on how it could be used could be useful. For example: provide use cases, questions it could help answer, etc.
2. Briefly introduce the basics of electricity markets, e.g., a very simplified version of a market clearing problem, its objectives and constraints. Similarly, introduce acronyms such as FCAS.
3. Explain how to interpret example results. What should be expected and how to know it worked? Could graphs be added to have a more visual result?
4. Specify the size of the data to download in the description of example 2 (this is only given in the comments). As this is really a lot of data, could it be possible to design another example with less data requirements?
5. Briefly explain the main steps of the dispatch procedure.
6. Be more specific when you mention: "Significant changes to the dispatch process are also likely to occur soon"
7. The text mentions "Version 1 will be a stable release and ongoing minor updates or patches will remain backwards compatible." The current version is 1.0.1
8. Link to a document explaining how the actual NEM works.
9. Explain how to use solvers and what criteria they should meet.
10. List "todos" you intend to implement in the future, possibly with priorities.
11. Move the long list of examples closer to where they would make sense, add more details in the description
12. Explain how to check the computation time.
13. Explain how to report an issue or get support.
14. List the dependencies.
15. Regarding the paper abstract, make it simpler for a wider audience and clarify the problems it can help solve and/or the questions it can help answer. Who is the target audience?
