### 2021-10-05 #3596 @noah80
@timtroendle
@nick-gorman

Sorry for the delay. After reviewing the software too, I reached similar conclusions to Robin.

1. The code is well-documented and readable.
2. Even after reading the documentation and the paper I still don't know what the software could be used for. Why would anyone want to model the dispatch?
3. The predefined examples show how to use the software, but don't explain the results. No graphs are generated. Could those be added to help the users better understand the results?
4. The research question for the software is not explained. Maybe add some other sources that explain why this software is used and what results can be generated with the software?

Right now with the given software it is rather difficult to judge if the software delivers the promised functionality and who could use the software for what purpose.
