+++ README.md
@@ -17,7 +17,7 @@ Operational dispatch is a important part of the NEM's decision making process. I
 
 1. As a reference implementation of the NEM's dispatch procedure. While the Australian Energy Market Operator (AEMO) has published several documents that describe aspects of the dispatch process, such as [FCAS Model in NEMDE](https://github.com/UNSW-CEEM/nempy/blob/master/docs/pdfs/FCAS%20Model%20in%20NEMDE.pdf), [Fast Start Unit Inflexibility Profile Model](https://github.com/UNSW-CEEM/nempy/blob/master/docs/pdfs/Fast_Start_Unit_Inflexibility_Profile_Model_October_2014.pdf), [Marginal Loss Factors](https://github.com/UNSW-CEEM/nempy/blob/master/docs/pdfs/Marginal%20Loss%20Factors%20for%20the%202020-21%20Financial%20year.pdf), [Schedule of Constraint Violation Penalty Factors](https://github.com/UNSW-CEEM/nempy/blob/master/docs/pdfs/Schedule%20of%20Constraint%20Violation%20Penalty%20factors.pdf), [Treatment of Loss Factors in the NEM](https://github.com/UNSW-CEEM/nempy/blob/master/docs/pdfs/Treatment_of_Loss_Factors_in_the_NEM.pdf), our experience developing Nempy has indicated that key implementation detail is often missing from the public available documentation. Through a process of testing various implementation options, where the documentation was not explicit, Nempy has been refined in an attempt to better reflect the actual dispatch procedure implementation. As result Nempy is a useful additional reference for analysts and modelers looking to understand the NEM's dispatch procedure.
 2. As a tool for studying the dispatch process itself. Models of any energy system or electricity market are necessarily simplification, however to improve model preformance it is often desirable to add additional detail. Nempy can be used to study the relative impact of different simplifications on modelling outcomes, and thus provide guidance on how model performance could be improved by adding additional detail. For example, someone working on an agent based model of the NEM might be interested in whether improving the agents bidding algorithm or improving the representation of network constraints will have a large impact on model peformance. The two figures in [section on Nempy's accuracy](https://nempy.readthedocs.io/en/latest/intro.html#accuracy) on the readthedocs page show how Nempy's accuracy is effected by removing key market features, and such analysis could be extended to look at the realtive impact of simiplify different market features or model inputs.
-3. As a building block in agent based market models. Agent based models can be used to study electricity market operation, and are particulary useful in capturing the competitive nature of these markets. In such models, agents must interact with a modelled environment, and a key part of that environment is the market dispatch process. Thus Nempy could be useful as building block to create agent based models of the NEM, and play role in answering the various questions about market operational outcomes to which these models are commonly applied.
+3. As a building block in agent based market models. Agent based models can be used to study electricity market operation, and are particulary useful in capturing the competitive nature of these markets. In such models, agents must interact with a modelled environment, and a key part of that environment is the market dispatch process. Thus, Nempy could be useful as building block to create agent based models of the NEM, and play role in answering the various questions about market operational outcomes to which these models are commonly applied.
 
 
 ## Documentation
