+++ paper.bib
@@ -69,6 +69,23 @@
   url = {https://github.com/martinblech/xmltodict}
 }
 
+@misc{nempro,
+  title = {NEMPRO},
+  author = {Gorman, Nicholas},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/UNSW-CEEM/NEMPRO}
+}
+
+@misc{nemde,
+  title = {NEM Dispatch Engine (NEMDE)},
+  author = {{Australian Energy Market Operator}},
+  year = {2021},
+  publisher = {Australian Energy Market Operator},
+  url = {https://aemo.com.au/en/energy-systems/market-it-systems/electricity-system-guides/nemde-queue-service}
+}
+
 @article{ringkjob2018review,
   title={A review of modelling tools for energy and electricity systems with large shares of variable renewables},
   author={Ringkj{\o}b, Hans-Kristian and Haugan, Peter M and Solbrekke, Ida Marie},
@@ -135,6 +152,16 @@
   publisher={Elsevier}
 }
 
+@article{ventosa,
+  title={Electricity market modeling trends},
+  author={Ventosa, Mariano and Baillo, Alvaro and Ramos, Andres and Rivier, Michel},
+  journal={Energy Policy},
+  volume={33},
+  pages={897--913},
+  year={2005},
+  publisher={Elsevier}
+}
+
 @misc{ANEMWorkingReport,
   title = {Energy Economics and Management Group Working Paper 2-2015},
   author = {P. Wild and W. P. Bell and J. Foster and M. Hewson},
@@ -166,4 +193,39 @@ changes},
   year = {2021},
   publisher = {Grattan Institute},
   url = {https://grattan.edu.au/wp-content/uploads/2021/04/Go-for-net-zero-Grattan-Report.pdf}
+}
+
+@misc{faststart
+  title = FAST-START INFLEXIBILITY PROFILE PROCESS DESCRIPTION,
+  author = {{Australian Energy Market Operator}},
+  year = {2014},
+  publisher = {{Australian Energy Market Operator}}
+}
+
+@misc{fcasmodel
+  title = FCAS Model in NEMDE
+  author = {{Australian Energy Market Operator}},
+  year = {2017},
+  publisher = {{Australian Energy Market Operator}}
+}
+
+@misc{lossfactors
+  title = Regions and Marginal Loss Factors: FY 2021-22
+  author = {{Australian Energy Market Operator}},
+  year = {2021},
+  publisher = {{Australian Energy Market Operator}}
+}
+
+@misc{constraintviolation
+  title = SCHEDULE OF CONSTRAINT VIOLATION PENALTY FACTORS
+  author = {{Australian Energy Market Operator}},
+  year = {2017},
+  publisher = {{Australian Energy Market Operator}}
+}
+
+@misc{treatmentlossfactors
+  title = TREATMENT OF LOSS FACTORS IN THE NATIONAL ELECTRICITY MARKET
+  author = {{Australian Energy Market Operator}},
+  year = {2012},
+  publisher = {{Australian Energy Market Operator}}
 }
