+++ paper.bib
@@ -195,36 +195,36 @@ changes},
   url = {https://grattan.edu.au/wp-content/uploads/2021/04/Go-for-net-zero-Grattan-Report.pdf}
 }
 
-@misc{faststart
-  title = FAST-START INFLEXIBILITY PROFILE PROCESS DESCRIPTION,
+@misc{faststart,
+  title = {FAST-START INFLEXIBILITY PROFILE PROCESS DESCRIPTION},
   author = {{Australian Energy Market Operator}},
   year = {2014},
   publisher = {{Australian Energy Market Operator}}
 }
 
-@misc{fcasmodel
-  title = FCAS Model in NEMDE
+@misc{fcasmodel,
+  title = {FCAS Model in NEMDE},
   author = {{Australian Energy Market Operator}},
   year = {2017},
   publisher = {{Australian Energy Market Operator}}
 }
 
-@misc{lossfactors
-  title = Regions and Marginal Loss Factors: FY 2021-22
+@misc{lossfactors,
+  title = {Regions and Marginal Loss Factors: FY 2021-22},
   author = {{Australian Energy Market Operator}},
   year = {2021},
   publisher = {{Australian Energy Market Operator}}
 }
 
-@misc{constraintviolation
-  title = SCHEDULE OF CONSTRAINT VIOLATION PENALTY FACTORS
+@misc{constraintviolation,
+  title = {SCHEDULE OF CONSTRAINT VIOLATION PENALTY FACTORS},
   author = {{Australian Energy Market Operator}},
   year = {2017},
   publisher = {{Australian Energy Market Operator}}
 }
 
-@misc{treatmentlossfactors
-  title = TREATMENT OF LOSS FACTORS IN THE NATIONAL ELECTRICITY MARKET
+@misc{treatmentlossfactors,
+  title = {TREATMENT OF LOSS FACTORS IN THE NATIONAL ELECTRICITY MARKET},
   author = {{Australian Energy Market Operator}},
   year = {2012},
   publisher = {{Australian Energy Market Operator}}
