### 2021-10-17 d54e9a5
add further explanations to examples
