### 2021-10-18 3c5a1b7
add more detail and charts to examples
