### 2021-10-18 eeaf015
modify xml manager populate, and add download warnings
