+++ docs/source/examples.rst
@@ -78,7 +78,8 @@ unit maximum capacity constraints and interconnector models, all sourced from hi
 
 *Results from example: for the QLD region a reasonable fit between modelled prices and historical prices is obtained.*
 
-.. warning:: Warning this script downloads approximately 8.5 GB of data from AEMO.
+.. warning:: Warning this script downloads approximately 8.5 GB of data from AEMO. The download_inputs flag can be set
+to false to stop the script re-downloading data for subsequent runs.
 
 .. literalinclude:: ../../examples/recreating_historical_dispatch.py
     :linenos:
@@ -96,7 +97,8 @@ conditions of the next interval, rather the historical initial conditions are al
 
 *Results from example: for the QLD region a very close fit between modelled prices and historical prices is obtained.*
 
-.. warning:: Warning this script downloads approximately 8.5 GB of data from AEMO.
+.. warning:: Warning this script downloads approximately 8.5 GB of data from AEMO. The download_inputs flag can be set
+to false to stop the script re-downloading data for subsequent runs.
 
 .. literalinclude:: ../../examples/all_features_example.py
     :linenos:
@@ -111,7 +113,8 @@ would need to be re-calculated based on the dynamic system state. Similarly, usi
 some what problematic as participants also dynamically change their bids based on market conditions. However, for the sake
 of demonstrating how Nempy can be used to create time sequential models, historical bids are used in this example.
 
-.. warning:: Warning this script downloads approximately 8.5 GB of data from AEMO.
+.. warning:: Warning this script downloads approximately 8.5 GB of data from AEMO. The download_inputs flag can be set
+to false to stop the script re-downloading data for subsequent runs.
 
 .. literalinclude:: ../../examples/time_sequential.py
     :linenos:
