### 2021-10-18 8ef96a8
edit paper and read the docs
