+++ paper.md
@@ -32,7 +32,7 @@ Nempy is a python package for modelling the dispatch procedure of the Australian
 Electricity markets are a way of co-ordinating the supply of electricity by private firms. The NEM is a gross pool spot 
 market that operates on 5 min dispatch basis. Described simply, this means all generators wishing to sell electricity 
 must bid into the market every 5 minutes, market clearing proceeds by calculating the cheapest combination of generator 
-dispatch to meet forecast demand at the end of 5 the minute dispatch interval. The price of electricity is set as the 
+operating levels to meet forecast demand at the end of 5 the minute dispatch interval. The price of electricity is set as the 
 marginal cost of generation, which, under a simple market formulation, would be the cost of the next generation bid to be 
 dispatched if demand for electricity were to increase. Real-world formulations require significant additional adjustment 
 in order to manage the technical complexity of securely and reliably operating an electricity grid. For example, in the 
@@ -41,7 +41,7 @@ been integrated into the market dispatch procedure are the Frequency Control Anc
 markets generators compete to provide the ability to rapidly change generation levels in order to stabilise the grid frequency. 
 Nempy is flexible in that it allows for formulation of very simple market models, for the formulation of market models 
 of near real-world complexity, and at the various levels of intermediate complexity. Simple models can be constructed 
-using just generator bids and electricity demand, so called bidstack models. More complete models can be constructed by 
+using just generator bids and electricity demand, so called bid stack models. More complete models can be constructed by 
 using the inbuilt features to create multiple market regions, ramp rate limits, loss factors, FCAS markets, FCAS trapezium 
 constraints, dynamic interconnector loss models, generic constraints and fast start dispatch inflexibility profiles. 
 Outputs include market clearing prices, generator and scheduled load dispatch targets, FCAS enablement levels, unit FCAS 
@@ -76,10 +76,22 @@ contexts. Some potential use case are outlined below:
 
 1. As a tool for studying the dispatch process itself. Models of any energy system or electricity market are necessarily 
 simplifications, however to improve model performance it is often desirable to add additional detail. Nempy can be used 
-to study the relative impact of different simplifications on modelling outcomes, and thus provide guidance on how model 
-performance could be improved by adding additional detail. For example, someone working on an operational model of the 
-NEM might be interested in which types of constraints have the biggest impact on dispatch outcomes, and thus should be 
-given priority for implementation.
+to study the impact of different simplifications on modelling outcomes, and thus provide guidance on how model 
+performance could be improved by adding additional detail. fig shows a simple example of such an analysis. The price
+results from the New South Wales region for 1000 randomly selected intervals in the 2019 calender year are shown. When
+Nempy is configured with a full set of market features price results closely match historical prices. When the FCAS 
+markets and generic constraints (network and security) are removed from the model results differ significantly. Resorting
+the results of the simpler market model, we can see that both models produce a similar number of medianly priced 
+intervals. However, the highest and lowest priced intervals of the simpler model are significantly lower. The average
+historical price is 81.4 $/MWh, the average price of the full featured model is 81.3 $/MWh, and the average price of the 
+simpler model is 75 $/MWh. The close match between the results of the full featured model and historical prices allows 
+for the attribution of the deviation of the simpler model explicitly to the simplification that have been made.  
+
+![Dispatch price results from the New South Wales region for 1000 randomly selected intervals in the 2019 calender year.
+  The historical prices, prior to scaling or capping are also shown for comparison. Results from two Nempy models are
+  shown, one with a full set of dispatch features, and one without FCAS markets or generic constraints (network and 
+  security constraints). For the simpler model price results are shown both historical price order and resorted.\label{fig:example}](plot.svg)
[…]
