+++ docs/source/intro.rst
@@ -1,8 +1,7 @@
 Introduction
 ============
 Nempy is an open-source python package that can be used to model the dispatch procedure of the Australian National
-Electricity Market (NEM). We believe nempy will be a valuable tool for academic researchers and industry analysts
-exploring a wide variety of questions. The dispatch process is at the core of many market modelling projects. As the
+Electricity Market (NEM). The dispatch process is at the core of many market modelling projects. As the
 NEM evolves, constraints and ancillary service markets are becoming increasingly important in shaping dispatch outcomes.
 Significant changes to the dispatch process are also likely to occur soon. Nempy allows users to easily configure a
 dispatch model to fit the relevant research question. Furthermore, if extra functionality is needed, the python
@@ -14,6 +13,34 @@ Nempy is feature rich, flexible, can recreate historical dispatch with a high de
 documentation and has planned support until mid-2023.
 
 Find nempy and more information at https://github.com/UNSW-CEEM/nempy.
+
+Dispatch Procedure Outline
+--------------------------
+The main step of the dispatch procedure is the construction and solving of a mixed integer linear problem (MIP) to find the
+least cost set of dispatch levels for generators and scheduled loads. Note, as the objective is technically to maximise the
+value of trade in the market, loads are treated as a negative costs. The construction of the MIP as implemented by
+Nempy proceeds roughly as follows:
+
+1. Bids from generators and loads are preprocessed, some FCAS bids are excluded if they do not meet a set of inclusion
+criteria set out by AEMO (:download:`FCAS Model in NEMDE <../../docs/pdfs/FCAS Model in NEMDE.pdf>`).
+2. For each bid from a generator a decision variable in the MIP is created, the cost of the variable is the bid price
+submitted by the generator, but is adjusted by a loss factor if one is provided.
+3. For each market region a constraint forcing generation to equal demand is created.
+4. The rest of the market features are implemented as additional variables and/or constraints in the MIP, for example:
+   - unit ramp rates are converted to a set MW ramp that units can achieve over the dispatch interval, and the sum of a
+     unit's dispatch is limited by this MW value
+   - interconnectors are formulated as additional decision variables that link the supply equals demand constraints
+     of the interconnected regions, and are combined with constraints sets that calculate interconnector losses as a
+     function of power flow
+5. The MIP is solved to determined interconnector flows and dispatch targets, the MIP is then converted to a linear
+   problem, and re-solved, such that market prices can be determined from constraint shadow prices.
+
+Other steps in the dispatch procedure that are not implemented by Nempy are:
+1. The preprocessing step to calculate of FCAS market and network constraint right hand side values (right hand side
+   values need to be provided as inputs to Nempy)
+2. Multiple re-runs of the optimisation to the operational settings for DC link between mainland synchronous region and
+   the Tasmainian synchronous region
+
 
 Features
 --------
