### 2021-10-18 1dc080c
docs intro minor updates
