### 2021-10-18 f94eae0
fix list in intro
