### 2021-10-18 21812b1
fix intro list
