### 2021-10-18 583c51d
update docs intro, change new xml populate functionality to keep it backward compatible
