+++ docs/source/intro.rst
@@ -3,28 +3,30 @@ Introduction
 Nempy is an open-source python package that can be used to model the dispatch procedure of the Australian National
 Electricity Market (NEM). The dispatch process is at the core of many market modelling projects. As the
 NEM evolves, constraints and ancillary service markets are becoming increasingly important in shaping dispatch outcomes.
-Significant changes to the dispatch process are also likely to occur soon. Nempy allows users to easily configure a
-dispatch model to fit the relevant research question. Furthermore, if extra functionality is needed, the python
-implementation, open-source licencing and planned ongoing support from developers make it possible to adapt nempy
-to your needs. Version 1 will be a stable release and ongoing minor updates or patches will remain backwards compatible.
-However, we are interested in receiving feedback to inform the ongoing maintenance, development and any future major updates.
+As part of the ongoing reform process significant changes to the dispatch process have also been proposed, for example,
+the introduction of an operating reserve market is being considered, as is the adoption of a dynamic intra-regional loss
+model. Nempy allows users to easily configure a dispatch model to fit the relevant research question. Furthermore, if
+extra functionality is needed, the python implementation, open-source licencing and planned ongoing support from developers
+make it possible to adapt Nempy to your needs. Version 1 will be a stable release and ongoing minor updates or patches
+will remain backwards compatible. However, we are interested in receiving feedback to inform the ongoing maintenance,
+development and any future major updates.
 
 Nempy is feature rich, flexible, can recreate historical dispatch with a high degree of accuracy, runs fast, has detailed
 documentation and has planned support until mid-2023.
 
-Find nempy and more information at https://github.com/UNSW-CEEM/nempy.
+The Nempy source code is on GitHub: https://github.com/UNSW-CEEM/nempy.
 
 Dispatch Procedure Outline
 --------------------------
-The task of the dispatch procedure is the construction and solving of a mixed integer linear problem (MIP) to find the
+The main task of the dispatch procedure is the construction and solving of a mixed integer linear problem (MIP) to find the
 least cost set of dispatch levels for generators and scheduled loads. Note, in this optimisation the dispatch of
 scheduled loads is treated as a negative cost, this makes the least cost optimisation equivalent to maximising the value of
 market trade. The construction of the MIP as implemented by Nempy proceeds roughly as follows:
 
 #. Bids from generators and loads are preprocessed, some FCAS bids are excluded if they do not meet a set of inclusion
    criteria set out by AEMO (:download:`FCAS Model in NEMDE <../../docs/pdfs/FCAS Model in NEMDE.pdf>`).
-#. For each bid from a generator and scheduled load a decision variable in the MIP is created, the cost of the
-   variable is the bid price, and the price is adjusted by a loss factor if one is provided.
+#. For each bid a decision variable in the MIP is created, the cost of the variable in the objective function is the bid
+   price, and the price is adjusted by a loss factor if one is provided.
 #. For each market region a constraint forcing generation to equal demand is created.
 #. The rest of the market features are implemented as additional variables and/or constraints in the MIP, for example:
 
@@ -73,9 +75,9 @@ the pandas' interface.
 
 Accuracy
 --------
-The accuracy with which nempy represents the NEM's dispatch process can be measured by re-creating historical dispatch results.
+The accuracy with which Nempy represents the NEM's dispatch process can be measured by re-creating historical dispatch results.
 This is done for a given dispatch interval by downloading the relevant historical inputs such as unit initial operating levels,
-bids and generic constraints, processing these inputs so they are compatible with the nempy SpotMarket class, and finally
+bids and generic constraints, processing these inputs so they are compatible with the Nempy SpotMarket class, and finally
[…]
