### 2021-10-18 9e10b96
fix example page formatting
