+++ docs/source/intro.rst
@@ -113,7 +113,7 @@ models by sourcing inputs from different data storage formats.
 
 Notes:
 
-- Information on solvers is provided is provided in the `spotmarket`_ documentation of the SpotMarket class.
+- Information on solvers is provided is provided in the `examples`_ documentation of the SpotMarket class.
 - The total runtime was calculated using the python time module and measuring the time taken from the loading of inputs
   to the extraction of results from the model. The runtime of different sub-process, i.e. loading of the XML file, was
   measure by inserting timing code into the Nempy source code where required.
