### 2021-10-18 420b425
fix links and xmlcache docs
