+++ docs/source/examples.rst
@@ -1,4 +1,4 @@
-.. _examples:
+.. _examples1:
 
 Examples
 ====================
