### 2021-10-19 42c38fd
delete travis yml, update project build setup, increment minor version number
