+++ .travis.yml
@@ -1,53 +1,0 @@
-language: python
-jobs:
-  include:
-    - name: "Linux python 3.6"
-      language: python
-      python:
-        - "3.6"
-    - name: "Linux python 3.6 oldest dependencies"
-      language: python
-      python:
-        - "3.6"
-      install:
-        - pip3 install --upgrade pip
-        - pip install pandas==1.0.0
-        - pip install mip==1.11.0
-        - pip install requests==2.0.0
-        - pip install e .
-        - pip install pytest
-        - pip show pandas
-        - pip show mip
-        - pip show xmltodict
-        - pip show requests
-    - name: "Linux python 3.7"
-      language: python
-      python:
-        - "3.7"
-    - name: "Linux python 3.8"
-      language: python
-      python:
-        - "3.8"
-    - name: "macOS"
-      os: osx
-      osx_image: xcode11.2
-      language: shell
-    - name: "Windows"
-      os: windows
-      language: shell
-      before_install:
-        - choco install python --version 3.8.0
-        - python -m pip install --upgrade pip
-      env: PATH=/c/Python38:/c/Python38/Scripts:$PATH
-# command to install dependencies
-install:
-  - pip3 install --upgrade pip
-  - pip install e .
-  - pip install pytest
-  - pip show pandas
-  - pip show mip
-  - pip show xmltodict
-  - pip show requests
-# command to run tests
-script:
-  - python3 -m pytest || python -m pytest
