+++ README.md
@@ -13,7 +13,7 @@ that you can start simple, like in the example below, and grow the complexity of
 ramping constraints, interconnectors, FCAS markets and more. 
 
 ## Motivation
-Operational dispatch is a important part of the NEM's decision making process. It allows information from generation operators, dispatchable or flexible loads, network operators, and the market operator to be simulatenous considered on a near real time basis in order to minimize costs while ensuring secure and reliable supply, and it creates market signals for longer term decisions. The motivation for developing Nempy was to create a re-usable model of this process 
+Operational dispatch is a important part of the NEM's decision making process. It allows information from generation operators, dispatchable or flexible loads, network operators, and the market operator to be simulatenous considered on a near real time basis in order to minimize costs while ensuring secure and reliable supply, and it creates market signals for longer term decisions. The motivation for developing Nempy was to create a re-usable model of this process.
 
 ## Documentation
 Find it on [readthedocs](https://nempy.readthedocs.io/en/latest/)
@@ -22,10 +22,10 @@ Find it on [readthedocs](https://nempy.readthedocs.io/en/latest/)
 Nempy is open-source and we welcome all forms of community engagment, some more info is provided below.
 
 ### Support
-You can seek suport for using Nempy using the discussion tab, checking the issues register, or by contacting Nick directly.
+You can seek suport for using Nempy using the discussion tab on GitHub (https://github.com/UNSW-CEEM/nempy/discussions), checking the issues register (https://github.com/UNSW-CEEM/nempy/issues), or by contacting Nick directly n.gorman<at>unsw.edu.au.
 
 ### Reporting issues
-Issues with Nempy can be reported via the issues register, issues submissions do not need to adhere to any particular format.
+Issues with Nempy can be reported via the issues register (https://github.com/UNSW-CEEM/nempy/issues), issues submissions do not need to adhere to any particular format.
 
 ### Contributing
 Contributions via pull requests are welcome. Contributions should; follow the PEP8 style guide (with exception of line length up to 120 rather than 80), ensure that all existing automated tests continue to pass (unless you are explicitly changing intended behavour, please highlight this in your pull request description), implement automated tests for new features, and provided doc strings for public interfaces. 
@@ -34,7 +34,7 @@ Contributions via pull requests are welcome. Contributions should; follow the PE
 This project is being lead by Nick Gorman, a PhD candidate at the University for New South Wales and the Collaboration 
 on Energy and Environmental Markets (CEEM). As part the project we hope to engage with and support prospective users of
 the software. Feel welcome to get in touch if you have any questions, want to provide feed back, have a feature request,
-are interested in collaborating or just want to discuss this project. You can contact Nick via n.gorman<@>unsw.edu.au.      
+are interested in collaborating or just want to discuss this project. You can contact Nick via n.gorman<at>unsw.edu.au.      
 
 ## Installation
 Installing nempy to use in your project is easy.
