+++ examples/all_features_example.py
@@ -190,39 +190,39 @@ fig.show()
 
 con.close()
 print(outputs)
-#                   time region      price
-# 0  2019/01/01 12:00:00   NSW1  91.870167
-# 1  2019/01/01 12:00:00   QLD1  76.190796
-# 2  2019/01/01 12:00:00    SA1  86.899534
-# 3  2019/01/01 12:00:00   TAS1  89.805037
-# 4  2019/01/01 12:00:00   VIC1  84.984255
-# 0  2019/01/01 12:05:00   NSW1  91.870496
-# 1  2019/01/01 12:05:00   QLD1  64.991736
-# 2  2019/01/01 12:05:00    SA1  87.462599
-# 3  2019/01/01 12:05:00   TAS1  90.178036
-# 4  2019/01/01 12:05:00   VIC1  85.556009
-# 0  2019/01/01 12:10:00   NSW1  91.870496
-# 1  2019/01/01 12:10:00   QLD1  64.991736
-# 2  2019/01/01 12:10:00    SA1  86.868556
-# 3  2019/01/01 12:10:00   TAS1  89.983716
-# 4  2019/01/01 12:10:00   VIC1  84.936150
-# 0  2019/01/01 12:15:00   NSW1  91.870496
-# 1  2019/01/01 12:15:00   QLD1  64.776456
-# 2  2019/01/01 12:15:00    SA1  86.844540
-# 3  2019/01/01 12:15:00   TAS1  89.582288
-# 4  2019/01/01 12:15:00   VIC1  84.990796
-# 0  2019/01/01 12:20:00   NSW1  91.870496
-# 1  2019/01/01 12:20:00   QLD1  64.776456
-# 2  2019/01/01 12:20:00    SA1  87.496112
-# 3  2019/01/01 12:20:00   TAS1  90.291144
-# 4  2019/01/01 12:20:00   VIC1  85.594840
-# 0  2019/01/01 12:25:00   NSW1  91.870167
-# 1  2019/01/01 12:25:00   QLD1  64.991736
-# 2  2019/01/01 12:25:00    SA1  87.519993
-# 3  2019/01/01 12:25:00   TAS1  90.488064
-# 4  2019/01/01 12:25:00   VIC1  85.630617
-# 0  2019/01/01 12:30:00   NSW1  91.870496
-# 1  2019/01/01 12:30:00   QLD1  64.991736
-# 2  2019/01/01 12:30:00    SA1  87.462000
-# 3  2019/01/01 12:30:00   TAS1  90.196284
-# 4  2019/01/01 12:30:00   VIC1  85.573321
+#                   time region      price       ROP
+# 0  2019/01/01 12:00:00   NSW1  91.870167  91.87000
+# 1  2019/01/01 12:00:00   QLD1  76.190796  76.19066
+# 2  2019/01/01 12:00:00    SA1  86.899534  86.89938
+# 3  2019/01/01 12:00:00   TAS1  89.805037  89.70523
+# 4  2019/01/01 12:00:00   VIC1  84.984255  84.98410
+# 0  2019/01/01 12:05:00   NSW1  91.870496  91.87000
+# 1  2019/01/01 12:05:00   QLD1  64.991736  64.99000
+# 2  2019/01/01 12:05:00    SA1  87.462599  87.46213
+# 3  2019/01/01 12:05:00   TAS1  90.178036  90.08096
+# 4  2019/01/01 12:05:00   VIC1  85.556009  85.55555
+# 0  2019/01/01 12:10:00   NSW1  91.870496  91.87000
+# 1  2019/01/01 12:10:00   QLD1  64.991736  64.99000
[…]
