+++ paper.md
@@ -38,7 +38,7 @@ dispatched if demand for electricity were to increase. Real-world formulations r
 in order to manage the technical complexity of securely and reliably operating an electricity grid. For example, in the 
 case of the NEM additional markets for ancillary services have been introduced. One set of ancillary markets that have 
 been integrated into the market dispatch procedure are the Frequency Control Ancillary Services (FCAS) markets. In these 
-markets generators compete to provide the ability to rapidly change generation levels in order to stabilise the grid frequency. 
+markets generators compete to provide the ability to rapidly change generation levels in order to control the grid frequency. 
 Nempy is flexible in that it allows for formulation of very simple market models, for the formulation of market models 
 of near real-world complexity, and at the various levels of intermediate complexity. Simple models can be constructed 
 using just generator bids and electricity demand, so called bid stack models. More complete models can be constructed by 
@@ -80,7 +80,7 @@ to study the impact of different simplifications on modelling outcomes, and thus
 performance could be improved by adding additional detail. Figure 1 shows a simple example of such an analysis. The price
 results from the New South Wales region for 1000 randomly selected intervals in the 2019 calender year are shown. When
 Nempy is configured with a full set of market features price results closely match historical prices. When the FCAS 
-markets and generic constraints (network and security) are removed from the model results differ significantly. Resorting
+markets and generic constraints (network and security) are removed from the model, results differ significantly. Resorting
 the results of the simpler market model, we can see that both models produce a similar number of medianly priced 
 intervals. However, the highest and lowest priced intervals of the simpler model are significantly lower. The average
 historical price is 81.4 $/MWh, the average price of the full featured model is 81.3 $/MWh, and the average price of the 
@@ -88,9 +88,9 @@ simpler model is 75 $/MWh. The close match between the results of the full featu
 for the attribution of the deviation of the simpler model explicitly to the simplification that have been made.  
 
 ![Dispatch price results from the New South Wales region for 1000 randomly selected intervals in the 2019 calender year.
-  The historical prices, prior to scaling or capping are also shown for comparison. Results from two Nempy models are
+  The historical prices, prior to scaling or capping, are also shown for comparison. Results from two Nempy models are
   shown, one with a full set of dispatch features, and one without FCAS markets or generic constraints (network and 
-  security constraints). For the simpler model price results are shown both historical price order and resorted.\label{fig:example}](plot.png)
+  security constraints). For the simpler model price results are shown both in historical price order and resorted.\label{fig:example}](plot.png)
 
 2. As a building block in agent based market models. Agent based models can be used to study electricity market 
 operation, and are particularly useful in modelling both the competitive nature of electricity markets and their complex 
