### 2021-10-19 698cf00
edits to joss paper, update docs
