### 2021-10-19 7cd1075
added links to nem fact sheet
