### 2021-10-20 0f0e8d8
update readthedocs intro and paper
