### 2021-10-21 b3e68ef
typo fix, figure caption, figure legend
