### 2021-10-21 #3596 @nick-gorman
Hi @robinroche and @noah, thanks again for reviewing and your help comments. I've drafted a new version of the paper, made some changes to the documentation, and released a new version of the software with a minor update to try and address them.

Bellow I've tried highlight specifically where and how I've address each of your comments, please ask me to clarify if needed.

## @robinroche comments:

1. While I understand the general purpose of the code, additional information on how it could be used could be useful. For example: provide use cases, questions it could help answer, etc.

I've added significantly to the statement of need section, lines 60 - 117. I've outlined 4 distinct use cases, and provided a simple example analysis that briefly demonstrates one of the use cases.

2. Briefly introduce the basics of electricity markets, e.g., a very simplified version of a market clearing problem, its objectives and constraints. Similarly, introduce acronyms such as FCAS

I've edited the summary section of the paper to address this, lines 10 - 13. I've introduced electricity markets, the way the NEM spot market works simply, and how it needs more complex features like FCAS in real life.

3. Explain how to interpret example results. What should be expected and how to know it worked? Could graphs be added to have a more visual result?

I have edited the examples in the documentation to address this. For examples 1 - 5 I have added additional inline comments to explain the results. For examples 6 and 7 I've added graphs to make interpreting the results easier.

https://nempy.readthedocs.io/en/latest/examples.html#examples

4. Specify the size of the data to download in the description of example 2 (this is only given in the comments). As this is really a lot of data, could it be possible to design another example with less data requirements?

For the 2nd example in the read me and examples 6 - 8 in the documentation I've added in text warnings about the volume of data to be downloaded.

I've also published a new version of the software that allows the historical data xml cache to be built on a daily rather than monthly basis. This reduces the data downloaded for the examples from 65GB to 8.5GB.

https://github.com/UNSW-CEEM/nempy/releases/tag/v1.1.0

5. Briefly explain the main steps of the dispatch procedure.

I've added a section to the documentation to outline the steps of the dispatch procedure.

https://nempy.readthedocs.io/en/latest/intro.html#dispatch-procedure-outline

6. Be more specific when you mention: "Significant changes to the dispatch process are also likely to occur soon"

I've added additional detail:

"As part of the ongoing reform process significant changes to the dispatch process have also been proposed, for example, the introduction of an operating reserve market is being considered, as is the adoption of a dynamic intra-regional loss model."

https://nempy.readthedocs.io/en/latest/intro.html#introduction

7. The text mentions "Version 1 will be a stable release and ongoing minor updates or patches will remain backwards compatible." The current version is 1.0.1

This is standard semantic versioning, it's still the major version 1, the second and third numbers just get updated to indicate minor updates (but backwards compatible) and patches respectively.

I've also tried to update the text to make this clearer without being verbose in the intro.

"Version 1 will be a stable release and ongoing minor updates or patches will remain backwards compatible. Currently the latest minor release is v1.1.0."

8. Link to a document explaining how the actual NEM works.

I've added links to a description of the NEM in both the readme and the readthedocs page:

https://github.com/UNSW-CEEM/nempy#introduction
https://nempy.readthedocs.io/en/latest/intro.html#introduction

Also, the fourth use case in the paper now explains that there is no complete description of the dispatch process provided by AEMO, and references several documents addressing aspects of the dispatch process.

9. Explain how to use solvers and what criteria they should meet.

The Run-time section (https://nempy.readthedocs.io/en/latest/intro.html#run-time) now links to the reference documentation on solvers.

10. List "todos" you intend to implement in the future, possibly with priorities.

The Ongoing Work section (updated) provides this information

https://nempy.readthedocs.io/en/latest/intro.html#ongoing-work

11. Move the long list of examples closer to where they would make sense, add more details in the description

Not sure exactly what you mean by moving the examples, there was an error where they were showing up in the top level menu, is this what you meant?

Some more detail has been added to the descriptions.

12. Explain how to check the computation time.

A brief explanation has been added.

"The total runtime was calculated using the python time module and measuring the time taken from the loading of inputs to the extraction of results from the model. The runtime of different sub-process, i.e. loading of the XML file, was measured by inserting timing code into the Nempy source code where required."

https://nempy.readthedocs.io/en/latest/intro.html#run-time

13. Explain how to report an issue or get support.

This is explained in the readme on GitHub, I've added links to the the issues register and discussion board to make this clearer.

https://github.com/UNSW-CEEM/nempy#community

14. List the dependencies.

Added this section to the documentation

https://nempy.readthedocs.io/en/latest/intro.html#dependencies

15. Regarding the paper abstract, make it simpler for a wider audience and clarify the problems it can help solve and/or the questions it can help answer. Who is the target audience?

Lines 10 - 17 of the paper should now provide a simpler intro for the wider audience. Lines 17 - 26 aim to softly introduce some of the additional complexity. Lines 60 - 117 deal specifically with use cases and example question. Lines 61 and 62 say who the target audience is.

I've also added a section to the documentation that talks about target users and uses cases:
https://nempy.readthedocs.io/en/latest/intro.html#example-use-cases

## @noah80 comments:

1. The code is well-documented and readable.

2. Even after reading the documentation and the paper I still don't know what the software could be used for. Why would anyone want to model the dispatch?

I've added significantly to the statement of need section, lines 60 - 117. I've outlined 4 distinct use cases, and provided a simple example analysis that briefly demonstrates one of the use cases.

3. The predefined examples show how to use the software, but don't explain the results. No graphs are generated. Could those be added to help the users better understand the results?

I have edited the examples in the documentation to address this. For examples 1 - 5 I have added additional inline comments to explain the results. For examples 6 and 7 I've added graphs to make interpreting the results easier.

https://nempy.readthedocs.io/en/latest/examples.html#examples

4. The research question for the software is not explained. Maybe add some other sources that explain why this software is used and what results can be generated with the software?

As for comment two I think lines 60 - 177 of the paper should now address this. I've also added a section on use cases to the documentation

https://nempy.readthedocs.io/en/latest/intro.html#example-use-cases
