### 2021-10-25 2a43b4f
add max capacity to DUDETAIL
