### 2021-10-28 4ef2e6a
fix ocdr conditions
