+++ nempy/markets.py
@@ -2812,7 +2812,9 @@ class SpotMarket:
                 self._market_constraints_rhs_and_type[constraint_group]['price'] = \
                     self._market_constraints_rhs_and_type[constraint_group]['constraint_id'].map(prices)
 
-        if 'generic_deficit' in self._decision_variables and allow_over_constrained_dispatch_re_run:
+        if ('generic_deficit' in self._decision_variables or 'fcas_deficit' in self._decision_variables) \
+                and allow_over_constrained_dispatch_re_run:
+
             fcas_ceiling_price_violated = False
             if 'fcas' in self._market_constraints_rhs_and_type:
                 if self._market_constraints_rhs_and_type['fcas']['price'].max() >= fcas_market_ceiling_price:
@@ -2829,14 +2831,16 @@ class SpotMarket:
                     energy_floor_price_violated = True
 
             generic_cons_violated = False
-            if (self._decision_variables['generic_deficit']['value'].max() > 0.0001 or
-                    self._decision_variables['generic_deficit']['value'].min() < -0.0001):
-                generic_cons_violated = True
+            if 'generic_deficit' in self._decision_variables:
+                if (self._decision_variables['generic_deficit']['value'].max() > 0.0001 or
+                        self._decision_variables['generic_deficit']['value'].min() < -0.0001):
+                    generic_cons_violated = True
 
             fcas_cons_violated = False
-            if (self._decision_variables['fcas_deficit']['value'].max() > 0.0001 or
-                    self._decision_variables['fcas_deficit']['value'].min() < -0.0001):
-                fcas_cons_violated = True
+            if 'fcas_deficit' in self._decision_variables:
+                if (self._decision_variables['fcas_deficit']['value'].max() > 0.0001 or
+                        self._decision_variables['fcas_deficit']['value'].min() < -0.0001):
+                    fcas_cons_violated = True
 
             if ((fcas_ceiling_price_violated or energy_ceiling_price_violated or energy_floor_price_violated) and
                     (generic_cons_violated or fcas_cons_violated)):
