### 2021-11-23 UNSW-CEEM/nempy#4 @nick-gorman
trouble running unit tests
Hi @noah80, You reported having trouble running the unit tests, could you please provide some details?

- Have you been able to install nempy and run any of the examples?
- Have you been able to install the code as per the the instructions for development purposes? (https://github.com/UNSW-CEEM/nempy#installation)
- What if any error messages are you receiving?
