### 2021-11-23 #3596 @nick-gorman
> The PDF update is excellent and fulfills now all criteria. I couldn't get the tests running so far though.
>
> @nick-gorman : Could you contact me? Maybe I'm doing something wrong.

@noah80 Good to hear I've been able to address your comments.

I've opened an issue on the nempy repo regarding the unit tests, see https://github.com/UNSW-CEEM/nempy/issues/4 . Can you reply on this issue and provide details of your problem running the unit tests?

I don't have your email, but you can contact me at n.gorman - at - unsw.edu.au
