### 2021-11-24 UNSW-CEEM/nempy#4 @nick-gorman
Thanks @noah80,

Also, which version of python are you using?

Can you try updating to the newest version of pytest? currently nempy asks for 6.0.1 but maybe that doesn't work with all setups
