### 2021-11-25 UNSW-CEEM/nempy#4 @noah80
Python 3.10.0

Pytest 6.0.1

Any ideas what I could try? As I said it's a fresh anaconda installation with a fresh virtual enviroment.
