### 2021-11-25 UNSW-CEEM/nempy#4 @nick-gorman
OK, yep so this is an issue specific to Python version 3.10, see stackoverflow below:

https://stackoverflow.com/questions/69528842/an-error-while-trying-to-execute-tests-on-python-3-10-with-pytest

Try updating pytest to 6.2.5 or above.

If that works, I'll update the nempy requirements so that pytest version 6.2.5 gets installed when you run pip install e .[dev]
