### 2021-11-29 UNSW-CEEM/nempy#4 @noah80
Thanks! Now it works. So you need to add pytest 6.2.5 to your requirements?

Next problem:

(nempyenv) C:\work\nempy\nempy\examples>python all_features_example.py
Traceback (most recent call last):
  File "C:\work\nempy\nempy\examples\all_features_example.py", line 5, in <module>
    import plotly.graph_objects as go
ModuleNotFoundError: No module named 'plotly'

Looks like plotly is missing in your requirements?
