### 2021-11-29 UNSW-CEEM/nempy#5 @nick-gorman
Plotly is deliberately left out of the requirements because its not needed to run nempy.

It's just being used with nempy in this particular example.

You should be able to get the example running, with a simple `pip install plotly`
