### 2021-11-29 UNSW-CEEM/nempy#4 @nick-gorman
Excellent, yes, I'll update the requirements for the dev install.

I've created a separate issue with regards to the plotly error.
