### 2021-12-01 a9406b8
fix issue with pytest and python 3.1.0
