+++ .github/workflows/test.yml
@@ -71,6 +71,21 @@ jobs:
       - name: Run tests with pytest
         run: python -m pytest
 
+  test-dev-install-on-linux-python-three-one:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install Python 3
+        uses: actions/setup-python@v1
+        with:
+          python-version: 3.1.0
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install e .[dev]
+      - name: Run tests with pytest
+        run: python -m pytest
+
   test-on-windows-python-three-eight:
     runs-on: windows-latest
     steps:
