### 2021-12-01 b9b2277
fix pytest with python 3.10
