### 2021-12-01 a8be888
try and make python 3.10 run for tests
