+++ .github/workflows/test.yml
@@ -94,7 +94,7 @@ jobs:
       - name: Install Python 3
         uses: actions/setup-python@v1
         with:
-          python-version: 3.10
+          python-version: 3.10.0
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
@@ -110,7 +110,7 @@ jobs:
       - name: Install Python 3
         uses: actions/setup-python@v1
         with:
-          python-version: 3.10
+          python-version: 3.10.0
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
