### 2021-12-01 UNSW-CEEM/nempy#4 @nick-gorman
Updated dev install requirements in release v1.1.1

https://github.com/UNSW-CEEM/nempy/releases/tag/v1.1.1
