### 2021-12-06 UNSW-CEEM/nempy#5 @noah80
Yes, I'm aware how to install plotly manually, thanks. :)

My point was that my understanding of the review includes checking that the examples work. And if the attempt to run the examples fails from lacking packages then that's not optimal. Shouldn't the dev requirements at least include everything you need to run the examples?

kaleido is missing too from the requirements by the way.

One suggestion regarding the all_features_examples: Maybe add a check if the file is already downloaded so that the user doesn't have to manually edit the code?
