### 2021-12-06 333c645
specify dependencies for examples
