### 2021-12-06 8241f9c
fix example notes and warnings
