### 2021-12-07 UNSW-CEEM/nempy#5 @nick-gorman
Hi @noah80, I see your point about the examples not working out of the box.

I'm reluctant to fix the issue by adding plotly and kaleido to the dev requirements as this doesn't fix the problem for users who just do the simple pip install nempy, and adding extra requirements is more likely to lead to dependencies conflicts.

Instead I've now added inline comments to the examples detailing the extra requirements, and the versions to install. I've also added a note box to the documentation to highlight these extra dependencies. See link:
https://nempy.readthedocs.io/en/latest/examples.html#detailed-recreation-of-historical-dispatch

What do you think about this approach?

Regarding the check for files already downloaded, this has issues where the data might be partially downloaded, and the download gets interrupted, say due to network issues or by the use interrupting the script. In this case in a subsequent run of the example the check for data would pass, and then the script would fail later, potentially with a confusing error.

If the user has to explicitly skip the download I think there is less chance of this happening, and if it does the fix (resetting the flag to download the data) is simpler and more obvious to the user.

Ideally, nempy would include a more sophisticated data check that would manage all of this, but that's a bit complicated and I haven't had the chance to implement it.
