### 2021-12-07 #3596 @nick-gorman
> Still having some issues with running the examples in a clean enviroment due to missing dependencies in the package. This is not critical for the functionality but bad for the usability of your software in my opinion and I would suggest fixing this.

@noah80, good point, I've tried to address as per UNSW-CEEM/nempy#5
