### 2021-12-14 #3596 @noah80
I was able to run all examples now after manually installing the missing dependencies.

While it would be nice to automate that part or surround make the code sections in the examples with try/except, it all works and looks usable.

@timtroendle All my issues have been adressed.

@nick-gorman Thanks for the cooperation!
