### 2021-12-20 #3596 @timtroendle
Thanks a lot @noah80 and @robinroche for your reviews!

We can now move on with the acceptance process. I had a quick read of the paper and found three small issues:

1. There is a typo in line 15: "the end of 5 the minute dispatch".
2. There are typos in line 2 and 3 of the legend in figure 1: "gernic constraints".
3. Whedon spotted 5 missing DOIs.

@nick-gorman , could you at this point please:
- [ ] Fix above-mentioned issues, where possible
