### 2022-02-06 a5c5cfc
fix typos, add missing dois
