### 2022-02-07 8b9d186
fix minor testing bugs
