+++ nempy/historical_inputs/mms_db.py
@@ -824,8 +824,7 @@ class InputsBySettlementDate(_MultiDataSource):
         ...   'SETTLEMENTDATE': ['2019/01/01 11:55:00', '2019/01/01 12:00:00'],
         ...   'INITIALMW': [1.0, 2.0]})
 
-        >>> data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
-        2
+        >>> _ = data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
 
         When we call get_data the output is filtered by SETTLEMENTDATE.
 
@@ -888,8 +887,7 @@ class InputsByIntervalDateTime(_MultiDataSource):
         ...   'INTERVAL_DATETIME': ['2019/01/01 11:55:00', '2019/01/01 12:00:00'],
         ...   'INITIALMW': [1.0, 2.0]})
 
-        >>> data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
-        2
+        >>> _ = data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
 
         When we call get_data the output is filtered by INTERVAL_DATETIME.
 
@@ -957,8 +955,7 @@ class InputsByDay(_MultiDataSource):
         ...   'SETTLEMENTDATE': ['2019/01/01 00:00:00', '2019/01/02 00:00:00'],
         ...   'INITIALMW': [1.0, 2.0]})
 
-        >>> data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
-        2
+        >>> _ = data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
 
         When we call get_data the output is filtered by SETTLEMENTDATE and the results from the appropriate market
         day starting at 04:05:00 are retrieved. In the results below note when the output changes
@@ -1048,8 +1045,7 @@ class InputsStartAndEnd(_SingleDataSource):
         ...   'END_DATE': ['2019/01/02 00:00:00', '2019/01/03 00:00:00'],
         ...   'INITIALMW': [1.0, 2.0]})
 
-        >>> data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
-        2
+        >>> _ = data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
 
         When we call get_data the output is filtered by START_DATE and END_DATE.
 
@@ -1131,8 +1127,7 @@ class InputsByMatchDispatchConstraints(_AllHistDataSource):
         ...   'VERSIONNO': [1, 2, 2, 3],
         ...   'RHS': [1.0, 2.0, 2.0, 3.0]})
 
-        >>> data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
-        4
+        >>> _ = data.to_sql('EXAMPLE', con=con, if_exists='append', index=False)
 
         >>> data = pd.DataFrame({
         ...   'SETTLEMENTDATE' : ['2019/01/02 00:00:00', '2019/01/02 00:00:00', '2019/01/03 00:00:00',
@@ -1142,8 +1137,7 @@ class InputsByMatchDispatchConstraints(_AllHistDataSource):
         ...                              '2019/01/03 00:00:00'],
         ...   'GENCONID_VERSIONNO': [1, 2, 2, 3]})
 
-        >>> data.to_sql('DISPATCHCONSTRAINT', con=con, if_exists='append', index=False)
-        4
+        >>> _ = data.to_sql('DISPATCHCONSTRAINT', con=con, if_exists='append', index=False)
 
         When we call get_data the output is filtered by the contents of DISPATCHCONSTRAINT.
 
[…]
