### 2022-02-07 #3596 @nick-gorman
@timtroendle

I've been able to fix the issues you found.

The release tag of the software is v1.1.2

And the zendo doi is: 10.5281/zenodo.5989170
