### 2022-02-09 #3596 @kthyng
Hi @nick-gorman! Items for wrapping up:

1. I see your Zenodo archive and looks good, except there seems to be a "2" in the middle of your title?
2. Please check the capitalization in your references. For example, Grozev et al has "australia" uncapitalized. You can preserve capitalization by putting {} around characters or strings in your .bib file.
