+++ paper.bib
@@ -242,7 +242,7 @@ changes},
 }
 
 @misc{nemfactsheet,
-  title = {FACT SHEET: The National Electricity Market},
+  title = {{FACT SHEET: The National Electricity Market}},
   author = {{Australian Energy Market Operator}},
   year = {2020},
   publisher = {{Australian Energy Market Operator}},
