+++ paper.bib
@@ -1,6 +1,6 @@
 @software{reback2020pandas,
     author       = {{The pandas development team}},
-    title        = {pandas-dev/pandas: Pandas},
+    title        = {{pandas-dev/pandas: Pandas}},
     month        = feb,
     year         = 2020,
     publisher    = {Zenodo},
@@ -20,7 +20,7 @@
 }
 
 @Article{         harris2020array,
- title         = {Array programming with {NumPy}},
+ title         = {{Array programming with NumPy}},
  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
                  van der Walt and Ralf Gommers and Pauli Virtanen and David
                  Cournapeau and Eric Wieser and Julian Taylor and Sebastian
@@ -43,7 +43,7 @@
 }
 
 @misc{coin-orpython-mip,
-  title = {coin-or/python-mip},
+  title = {{coin-or/python-mip}},
   author = {H. G. Santos and T. A. M. Toffolo},
   year = {2021},
   publisher = {GitHub},
@@ -52,7 +52,7 @@
 }
 
 @misc{psf,
-  title = {Requests},
+  title = {{Requests}},
   author = {{Python Software Foundation}},
   year = {2021},
   publisher = {GitHub},
@@ -61,7 +61,7 @@
 }
 
 @misc{xmltodict,
-  title = {xmltodict},
+  title = {{xmltodict}},
   author = {M. Blech},
   year = {2021},
   publisher = {GitHub},
@@ -70,7 +70,7 @@
 }
 
 @misc{nempro,
-  title = {NEMPRO},
+  title = {{NEMPRO}},
   author = {Gorman, Nicholas},
   year = {2021},
   publisher = {GitHub},
@@ -79,7 +79,7 @@
 }
 
 @misc{nemde,
-  title = {NEM Dispatch Engine (NEMDE)},
+  title = {{NEM Dispatch Engine (NEMDE)}},
   author = {{Australian Energy Market Operator}},
   year = {2021},
   publisher = {Australian Energy Market Operator},
@@ -87,7 +87,7 @@
 }
 
 @article{ringkjob2018review,
-  title={A review of modelling tools for energy and electricity systems with large shares of variable renewables},
+  title={{A review of modelling tools for energy and electricity systems with large shares of variable renewables}},
   author={Ringkj{\o}b, Hans-Kristian and Haugan, Peter M and Solbrekke, Ida Marie},
   journal={Renewable and Sustainable Energy Reviews},
   volume={96},
@@ -98,7 +98,7 @@
 }
 
 @article{chang2021trends,
-  title={Trends in tools and approaches for modelling the energy transition},
+  title={{Trends in tools and approaches for modelling the energy transition}},
   author={Chang, Miguel and Thellufsen, Jakob Zink and Zakeri, Behnam and Pickering, Bryn and Pfenninger, Stefan and Lund, Henrik and {\O}stergaard, Poul Alberg},
   journal={Applied Energy},
   volume={290},
@@ -109,7 +109,7 @@
 }
 
 @article{fattahi2020systemic,
-  title={A systemic approach to analyze integrated energy system modeling tools: A review of national models},
+  title={{A systemic approach to analyze integrated energy system modeling tools: A review of national models}},
   author={Fattahi, Amirhossein and Sijm, Jos and Faaij, Andr{\'e}},
   journal={Renewable and Sustainable Energy Reviews},
   volume={133},
@@ -120,7 +120,7 @@
 }
 
 @inproceedings{grozev2005nemsim,
-  title={NEMSIM: Agent-based simulator for Australia's national electricity market},
[…]
