### 2022-02-09 #3596 @nick-gorman
Thanks @kthyng

Not sure how that 2 got there, or how I didn't see it. Good tip with the curly brackets.

These issues should be resolved.
