### 2022-02-10 #3596 @kthyng
@nick-gorman ok everything looks good! But, do you want me to  accept this submission from the `joss-paper` branch or were you going to merge it first? Either way is fine, I just need to know which.
