### 2022-02-10 #3596 @nick-gorman
@kthyng I've just merged joss-paper into master, so should be good to go from master
